requirements.txt
Flask>=2.3
pyyaml>=6.0
openai>=1.0.0
colorama>=0.4.6
